Create a dispatch switch in the raft server.
This follows the pattern of all information being carried
in the signals.

The signal is then matched in order to decided on its type
and then to process it accordingly. Again, while we tend to
still think in terms of "named" behaviours, the type is
actually the name that we are looking for.