WIP incomplete implementation of candidate to follower conversion.
The candidate must convert to a follower if it sees a heartbeat with the
same or higher term (i.e. a peer was elected as the leader).

However, this is difficult to arrange because currently monitors are of
type `val` and therefore can not modify their state. The monitor should
probably be changed to `iso` allowing the methods to be of type `ref`.